refactor implementation of FixedTrashText
