+++ components/src/hardware-sim/Deck/OT2Layers.tsx
@@ -35,22 +35,6 @@ export function FixedTrash(props: FixedTrashProps): JSX.Element {
     </g>
   )
 }
-
-export function FixedTrashText(): JSX.Element {
-  return (
-    <g id="fixedTrashText" transform="translate(291.5, 322.5) scale(1, -1)">
-      <path
-        d="M4.25736 18.3891C4.25736 18.5479 4.36322 18.6537 4.52197 18.6537H7.51185C7.6706 18.6537 7.77642 18.5479 7.77642 18.3891V3.73084C7.77642 3.625 7.82936 3.57208 7.93519 3.57208H11.3484C11.5071 3.57208 11.613 3.46625 11.613 3.30749V0.926183C11.613 0.767428 11.5071 0.661593 11.3484 0.661593H0.711879C0.553125 0.661593 0.447266 0.767428 0.447266 0.926183V3.30749C0.447266 3.46625 0.553125 3.57208 0.711879 3.57208H4.09862C4.20446 3.57208 4.25736 3.625 4.25736 3.73084V18.3891Z M24.9547 18.4156C25.0341 18.6008 25.1135 18.6537 25.2722 18.6537H28.7913C29.0029 18.6537 29.003 18.5214 28.9236 18.3891L25.4045 11.9067C25.1135 11.4304 24.9812 11.1394 24.9812 10.8748C24.9812 10.7689 25.0606 10.6631 25.2458 10.5043C27.0715 9.44599 27.5742 8.0172 27.5742 5.63589C27.5742 2.011 25.5368 0.661593 21.5151 0.661593H16.1439C15.9851 0.661593 15.8793 0.767428 15.8793 0.926183V18.3891C15.8793 18.5479 15.9851 18.6537 16.1439 18.6537H19.1338C19.2925 18.6537 19.3983 18.5479 19.3983 18.3891V12.039C19.3983 11.9067 19.4777 11.8538 19.61 11.8538H21.4622C21.6209 11.8802 21.6738 11.9596 21.8855 12.3565L24.9547 18.4156ZM20.88 9.47245H19.5835C19.4513 9.44599 19.3983 9.39307 19.3983 9.26077V3.38687C19.3983 3.28104 19.4513 3.22812 19.7423 3.22812H20.88C23.2613 3.22812 23.8964 3.83668 23.8964 6.21799C23.8964 8.57284 23.1555 9.47245 20.88 9.47245Z M42.7298 18.3891C42.7827 18.6008 42.8886 18.6537 43.0738 18.6537H46.4605C46.5928 18.6537 46.6457 18.6008 46.6457 18.495C46.6457 18.4685 46.6457 18.4156 46.6457 18.3891C46.0901 16.0607 43.1532 4.86858 42.0154 0.926183C41.9625 0.767428 41.8831 0.661593 41.6979 0.661593H36.6442C36.4855 0.661593 36.3797 0.767428 36.3267 0.926183C35.2155 4.60399 32.3314 15.1876 31.5377 18.3891C31.5112 18.4156 31.5112 18.4685 31.5112 18.495C31.5112 18.6008 31.5906 18.6537 31.7229 18.6537H34.7921C34.9509 18.6537 35.0567 18.6008 35.1361 18.3891L36.1415 14.288C36.168 14.1557 36.2738 14.1028 36.4061 14.1028H41.5656C41.6979 14.1028 41.7508 14.1557 41.7773 14.288L42.7298 18.3891ZM36.9088 11.4304C36.803 11.4304 36.7501 11.404 36.7501 11.2717C37.0411 9.92225 38.073 5.71526 38.6287 3.65146C38.6551 3.54563 38.7345 3.51917 39.0785 3.51917C39.4225 3.51917 39.5018 3.54563 39.5283 3.65146C40.031 5.71526 40.9835 9.92225 41.2746 11.2717C41.2746 11.404 41.2217 11.4304 41.1158 11.4304H36.9088Z M60.0047 3.7573C60.1105 3.7573 60.1634 3.67792 60.1634 3.59854L60.3222 1.29661C60.3222 1.21723 60.2957 1.1114 60.1899 1.05848C59.5548 0.793888 57.6234 0.529297 56.1152 0.529297C52.2786 0.529297 50.3207 1.53474 50.3207 5.08025C50.3207 6.93238 50.7969 8.14949 53.0459 9.84287C56.6708 12.5946 56.7767 13.0709 56.7767 14.0499C56.7767 15.5051 56.1152 15.9549 54.5541 15.9549C53.1782 15.9549 51.8024 15.8226 50.4794 15.5845C50.3471 15.5845 50.2942 15.6639 50.2942 15.7697L50.109 18.151C50.109 18.2568 50.1884 18.3362 50.3471 18.3627C51.379 18.6008 53.2312 18.8389 55.3214 18.8389C58.6553 18.8389 60.4545 17.4631 60.4545 14.0499C60.4545 11.4569 59.2903 10.2927 57.0677 8.73159C54.2895 6.77362 53.9455 6.29736 53.9455 5.239C53.9455 3.91605 54.5012 3.38687 56.0887 3.38687C57.2529 3.38687 59.0521 3.51917 60.0047 3.7573Z M75.0371 7.85844C75.0371 7.96428 74.9842 8.0172 74.8784 8.0172H69.4013C69.2955 8.0172 69.2426 7.96428 69.2426 7.85844V0.926183C69.2426 0.767428 69.1367 0.661593 68.978 0.661593H65.9881C65.8294 0.661593 65.7235 0.767428 65.7235 0.926183V18.3891C65.7235 18.5479 65.8294 18.6537 65.9881 18.6537H68.978C69.1367 18.6537 69.2426 18.5479 69.2426 18.3891V10.9806C69.2426 10.8748 69.2955 10.8219 69.4013 10.8219H74.8784C74.9842 10.8219 75.0371 10.8748 75.0371 10.9806V18.3891C75.0371 18.5479 75.143 18.6537 75.3017 18.6537H78.318C78.4768 18.6537 78.5826 18.5479 78.5826 18.3891V0.926183C78.5826 0.767428 78.4768 0.661593 78.318 0.661593H75.3017C75.143 0.661593 75.0371 0.767428 75.0371 0.926183V7.85844Z"
-        style={{
-          fill: '#C4C4C4',
-          width: '100%',
-          height: '100%',
-        }}
-      />
-    </g>
-  )
-}
-
 export function DoorStops(): JSX.Element {
   return (
     <g id="doorStops">
