+++ protocol-designer/src/molecules/HardwareOverlays/index.ts
@@ -1,0 +1,1 @@
+export { FixedTrashText } from './FixedTrashText'
