+++ protocol-designer/src/molecules/index.ts
@@ -3,6 +3,7 @@ export * from './CheckboxStepFormField'
 export * from './DropdownStepFormField'
 export * from './EndUserAgreementFooter'
 export * from './InputStepFormField'
+export * from './HardwareOverlays'
 export * from './LiquidButton'
 export * from './ToggleExpandStepFormField'
 export * from './ToggleStepFormField'
